Metrics: move Witness class to better package names
Part of #7788
