+++ logstash-core/src/main/java/org/logstash/instrument/witness/pipeline/ErrorWitness.java
@@ -1,4 +1,4 @@
-package org.logstash.instrument.witness;
+package org.logstash.instrument.witness.pipeline;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
@@ -6,11 +6,12 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
 import org.logstash.instrument.metrics.Metric;
 import org.logstash.instrument.metrics.gauge.TextGauge;
+import org.logstash.instrument.witness.MetricSerializer;
+import org.logstash.instrument.witness.SerializableWitness;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
-import java.nio.charset.StandardCharsets;
 
 /**
  * Witness for errors.
@@ -132,7 +133,7 @@ public class ErrorWitness implements SerializableWitness {
         /**
          * Gets the error message
          *
-         * @return the error message
+         * @return the error message. May be {@code null}
          */
         public String message() {
             return witness.message.getValue();
@@ -141,7 +142,7 @@ public class ErrorWitness implements SerializableWitness {
         /**
          * Gets the error stack/back trace
          *
-         * @return the backtrace as a String
+         * @return the backtrace as a String. May be {@code null}
          */
         public String backtrace() {
             return witness.backtrace.getValue();
