+++ logstash-core/src/main/java/org/logstash/instrument/witness/pipeline/PipelineWitness.java
@@ -1,9 +1,11 @@
-package org.logstash.instrument.witness;
+package org.logstash.instrument.witness.pipeline;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
+import org.logstash.instrument.witness.configuration.ConfigWitness;
+import org.logstash.instrument.witness.SerializableWitness;
 
 import java.io.IOException;
