+++ logstash-core/src/main/java/org/logstash/instrument/witness/pipeline/PluginsWitness.java
@@ -1,12 +1,12 @@
-package org.logstash.instrument.witness;
+package org.logstash.instrument.witness.pipeline;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
+import org.logstash.instrument.witness.SerializableWitness;
 
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
