+++ logstash-core/src/test/java/org/logstash/instrument/witness/pipeline/ErrorWitnessTest.java
@@ -1,8 +1,9 @@
-package org.logstash.instrument.witness;
+package org.logstash.instrument.witness.pipeline;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Before;
 import org.junit.Test;
+import org.logstash.instrument.witness.pipeline.ErrorWitness;
 
 import static org.assertj.core.api.Assertions.assertThat;
