+++ logstash-core/src/test/java/org/logstash/instrument/witness/pipeline/PluginsWitnessTest.java
@@ -1,9 +1,10 @@
-package org.logstash.instrument.witness;
+package org.logstash.instrument.witness.pipeline;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Before;
 import org.junit.Test;
+import org.logstash.instrument.witness.pipeline.PluginsWitness;
 
 import static org.assertj.core.api.Assertions.assertThat;
